+++ src/main/java/com/responseor/ootw/service/MemberServiceImpl.java
@@ -100,7 +100,7 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public void deleteMemberClothes(Long id) {
+    public void deleteMemberClothes(int id) {
         ClothesByTemp clothesByTemp = clothesByTempRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_CLOTHES));
 
         clothesByTempRepository.delete(clothesByTemp);
